Moved all event-dispatching stuff to its own module
